Adding more tests

+++ Zander.Domain/Remote/IRemoteServerApi.cs
@@ -6,7 +6,7 @@ using Zander.Domain.Entities;
 namespace Zander.Domain.Remote {
 	public interface IRemoteServerApi : IDisposable {
 		MasterChallengeResponse ChallengeMasterServer(MasterChallengeRequest request);
-		IEnumerable<IPEndPoint> GetServerList(string masterServerAddress, long challenge, short protocolVersion);
+		IEnumerable<IPEndPoint> GetServerList(string masterServerAddress, int challenge, short protocolVersion);
 		Server GetServerInfo(IPEndPoint serverEndpoint);
 	}
 }
